- move file to config folder - add general layout - make compoment names more specific - add layout
