+++ libs/common/ui/src/lib/common-ui-header/common-ui-header.component.html
@@ -1,6 +1,3 @@
 <header class="py-6 sticky top-0 z-10 bg-ugurtigu-base">
-  <nav class="max-w-[1280px] mx-auto flex justify-between px-4">
-    Lorem ipsum dolor sit amet, consectetur adipisicing elit. Obcaecati,
-    inventore.
-  </nav>
+  <nav class="max-w-[1280px] mx-auto flex justify-between px-4">header</nav>
 </header>
