--- libs/layout/feature/src/lib/footer/footer.component.html
@@ -1,1 +1,0 @@
-<p>footer works!</p>
