--- libs/layout/feature/src/lib/footer/footer.component.ts
@@ -1,11 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'ugurtigu-footer',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './footer.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class FooterComponent {}
