+++ libs/layout/feature/src/lib/layour-feature-footer/layout-feature-footer.component.html
@@ -1,0 +1,1 @@
+<footer class="mx-auto max-w-[80ch] py-12 fade-in px-4">footer</footer>
