+++ libs/layout/feature/src/lib/layout-feature-main/layout-feature-main.component.html
@@ -1,0 +1,1 @@
+<main class="max-w-[80ch] mx-auto px-4">main</main>
