+++ libs/layout/feature/src/lib/main-layout/main-layout.component.html
@@ -1,5 +1,7 @@
-<body class="bg-ugurtigu-base text-ugurtigu-contrast">
+<body class="bg-ugurtigu-base text-ugurtigu-contrast flex flex-col h-screen">
   <ugurtigu-layout-feature-header></ugurtigu-layout-feature-header>
-  <ugurtigu-main></ugurtigu-main>
-  <ugurtigu-footer></ugurtigu-footer>
+  <div class="flex-grow">
+    <ugurtigu-layout-feature-main></ugurtigu-layout-feature-main>
+  </div>
+  <ugurtigu-layout-feature-footer></ugurtigu-layout-feature-footer>
 </body>
