+++ libs/layout/feature/src/lib/main-layout/main-layout.component.ts
@@ -1,18 +1,17 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { FooterComponent } from '../footer/footer.component';
+import { LayoutFeatureFooterComponent } from '../layour-feature-footer/layout-feature-footer.component';
 import { LayoutFeatureHeaderComponent } from '../layout-feature-header/layout-feature-header.component';
-
-import { MainComponent } from '../main/main.component';
+import { LayoutFeatureMainComponent } from '../layout-feature-main/layout-feature-main.component';
 
 @Component({
   selector: 'ugurtigu-main-layout',
   standalone: true,
   imports: [
     CommonModule,
-    MainComponent,
+    LayoutFeatureMainComponent,
     LayoutFeatureHeaderComponent,
-    FooterComponent,
+    LayoutFeatureFooterComponent,
   ],
   templateUrl: './main-layout.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
