--- libs/layout/feature/src/lib/main/main.component.html
@@ -1,1 +1,0 @@
-<p>main works!</p>
