--- libs/layout/feature/src/lib/main/main.component.ts
@@ -1,11 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'ugurtigu-main',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './main.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class MainComponent {}
